requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.1
django-bootstrap-form==3.4
django-cors-headers==3.13.0
django-simple-history>=3.1.1
djangorestframework==3.13.1
numpy==1.22.4
Pillow==9.2.0
psycopg2-binary==2.9.3
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
boto3==1.24.46
